requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-celery==3.3.0
psycopg2-binary==2.9.9
redis==4.6.0